refactor: add root path 'admin'
feat: extend component to accept nav links array
