+++ packages/app/src/app/[locale]/(main)/admin/rights/RightsView.tsx
