+++ packages/app/src/app/[locale]/(main)/admin/rights/page.tsx
