+++ packages/app/src/app/[locale]/(main)/admin/translations/TranslationsView.tsx
