+++ packages/app/src/app/[locale]/(main)/admin/translations/page.tsx
